flask-backend/requirements.txt
Flask==2.2.2
Flask_Cors==3.0.10
numpy==1.23.2
opencv_python==4.6.0.66
python-consul==1.1.0
flask_marshmallow==0.14.0
Flask_SQLAlchemy==2.5.1
jose==1.0.0
Flask_Classful==0.14.2